+++ examples/tsibstn/main.tf
@@ -1,6 +1,10 @@
+provider "aws" {
+  region = "ap-southeast-1"
+}
+
 locals {
-  vpc_id               = "vpc-11aa22bb"
-  ami_owner_account_id = "1234567890"
+  vpc_id               = "vpc-11abbccdd"
+  ami_owner_account_id = "012345678"
   environment          = "staging"
   product_domain       = "tsi"
   service_name         = "tsibstn"
@@ -16,17 +20,4 @@ module "this" {
   ami_owner_account_id = local.ami_owner_account_id
   asg_capacity         = local.asg_capacity
   description          = "bastion for ${local.product_domain}"
-}
-
-# singleton, you'll only need 1 of this for your AWS account
-module "sesson_manager_config" {
-  source         = "git@github.com:traveloka/terraform-aws-session-manager-config.git?ref=v0.2.0"
-  environment    = local.environment
-  product_domain = local.product_domain
-}
-
-# this policy is required to enable your bastion VM to be ssm ready
-resource "aws_iam_role_policy_attachment" "bastion_policy" {
-  role       = module.this.instance_role_name
-  policy_arn = module.sesson_manager_config.iam_policy_arn
 }